conformance: move test into a new module
